requirements.txt
Alfred-PyWorkflow==2.0.0b3
